All changes published & visible
These are all the changes made to the original 1.02 source code so far.
This makes them all easily visible.

+++ CODE-mp/botlib/be_aas_reach.cpp
@@ -575,7 +575,7 @@ int AAS_FallDamageDistance(void)
 {
 	float maxzvelocity, gravity, t;
 
-	maxzvelocity = sqrt(30 * 10000);
+	maxzvelocity = sqrt((double)(30 * 10000));
 	gravity = aassettings.phys_gravity;
 	t = maxzvelocity / gravity;
 	return 0.5 * gravity * t * t;
